\name{richest}
\alias{richest}
\title{Richness estimator function}
\description{A function for estimating sample richnesses for samples with given sizes, to be taken from a population with given initial sample data}
\usage{
  richest(x,sampleSizes,method='Lijou',supports=0,pop=0)
}
\arguments{
  \item{x}{Initial sample data: a "richestInput" object, generated by one two methods: 
  1. a call to one of the functions: "readCPS", "readSPC", or
  2. generated from a data.frame "x" by the one of the commands: new("richestSPC",x), new("richestCPS",x)}
  \item{sampleSizes}{A vector of integer values indicating the sizes of samples for which to estimate sample richnesses.}
  \item{method}{An estimation method, i.e. one of the following character strings: "Lijou", "GT", or "PNPMLE".}
  \item{supports}{For the PNPMLE method only, an integer giving the number of support points to use when estimating the maximum-likelihood mixture distribution for the Poisson parameters.  Default: 0 means that the number of support points equals the number of distinct abundances in the sample.}
  \item{pop}{For the PNPMLE method only, a numeric value estimating the population richness.  Default: 0 causes the population richness to be given by the Chao lower bound.}
}
\references{ }
\author{Christopher Durden, Center for Genomics and Bioinformatics, Indiana University}
\keyword{richness}
\seealso{ readCPS }
\examples{
## read CPS data into cps and run Lijou method
cps = data(sargassoOTU)
est = richest(cps,sampleSizes=seq(1,10000,by=25),method="Lijou")
est = richest(cps,sampleSizes=seq(1,10000,by=25),method="PNPMLE")
est = richest(cps,sampleSizes=seq(1,1400,by=25),method="GT")
}
